+++ tests/ValueObjects/DefaultValuesBothVO.php
@@ -1,0 +1,27 @@
+<?php
+
+declare (strict_types = 1);
+
+namespace Tests\ValueObjects;
+
+use ReallifeKip\ImmutableBase\Attributes\Defaults;
+use ReallifeKip\ImmutableBase\Objects\ValueObject;
+
+readonly class DefaultValuesBothVO extends ValueObject
+{
+    #[Defaults(true)]
+    public ?bool $bool;
+    #[Defaults(1)]
+    public ?int $int;
+    #[Defaults([9, 9, 9])]
+    public ?array $array;
+
+    public static function defaultValues(): array
+    {
+        return [
+            'bool'  => false,
+            'int'   => 0,
+            'array' => [],
+        ];
+    }
+}