+++ tests/ValueObjects/UncacheableDefaultValueVO.php
@@ -1,0 +1,19 @@
+<?php
+
+declare (strict_types = 1);
+
+namespace Tests\ValueObjects;
+
+use ReallifeKip\ImmutableBase\Objects\ValueObject;
+
+readonly class UncacheableDefaultValueVO extends ValueObject
+{
+    public array $functions;
+
+    public static function defaultValues(): array
+    {
+        return [
+            'functions' => [function () {}],
+        ];
+    }
+}